Fixes disconnected vnd.docker.raw-streams output
Oftentimes our executions would seem to bail out. They would continue to run, but the stream would be severed and we could not see the output from the stream. Traced it to the data variable being overwritten, and thus the buffer not properly being updated for the next bucket.  There may still be an issue with the very last bucket still containing more than one decoded control code series, but this gets us most of the way there.